move all the serial stuff to dedicated public module
