+++ dnp3/examples/outstation_tcp_server.rs
@@ -1,11 +1,10 @@
 use dnp3::app::control::*;
 use dnp3::app::measurement::*;
 use dnp3::decode::*;
-use dnp3::link::{EndpointAddress, LinkErrorMode};
+use dnp3::link::*;
 use dnp3::outstation::database::*;
-use dnp3::outstation::tcp::{AddressFilter, TcpServer};
-use dnp3::outstation::OutstationConfig;
 use dnp3::outstation::*;
+use dnp3::tcp::*;
 
 use std::time::Duration;
