+++ dnp3/src/lib.rs
@@ -69,6 +69,8 @@ pub mod link;
 pub mod master;
 /// Types and traits specific to outstations
 pub mod outstation;
+/// Entry points and types for serial
+pub mod serial;
 /// Entry points and types for TCP
 pub mod tcp;
