+++ dnp3/src/tcp/outstation.rs
@@ -1,54 +1,12 @@
 use crate::outstation::task::OutstationTask;
 
 use crate::link::LinkErrorMode;
-use crate::outstation::config::OutstationConfig;
 use crate::outstation::database::EventBufferConfig;
-use crate::outstation::traits::{ControlHandler, OutstationApplication, OutstationInformation};
 use crate::outstation::OutstationHandle;
+use crate::outstation::*;
+use crate::tcp::{AddressFilter, FilterError};
 use crate::util::task::Shutdown;
 use tracing::Instrument;
-
-#[derive(Clone, Debug, PartialEq)]
-pub enum AddressFilter {
-    Any,
-    Exact(std::net::IpAddr),
-    AnyOf(std::collections::HashSet<std::net::IpAddr>),
-}
-
-impl AddressFilter {
-    pub(crate) fn matches(&self, addr: std::net::IpAddr) -> bool {
-        match self {
-            AddressFilter::Any => true,
-            AddressFilter::Exact(x) => *x == addr,
-            AddressFilter::AnyOf(set) => set.contains(&addr),
-        }
-    }
-
-    pub(crate) fn conflicts_with(&self, other: &AddressFilter) -> bool {
-        match self {
-            AddressFilter::Any => true,
-            AddressFilter::Exact(x) => other.matches(*x),
-            AddressFilter::AnyOf(set) => set.iter().any(|x| other.matches(*x)),
-        }
-    }
-}
-
-/// error type returned when a filter conflicts with another filter
-#[derive(Copy, Clone, Debug, PartialEq)]
-pub enum FilterError {
-    /// filter conflicts with an existing filter
-    Conflict,
-}
-
-impl std::error::Error for FilterError {}
-
-impl std::fmt::Display for FilterError {
-    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
-        match self {
-            FilterError::Conflict => f.write_str("filter conflicts with an existing filter"),
-        }
-    }
-}
 
 struct Outstation {
     filter: AddressFilter,
